requirements.txt
fastapi
pandas
numpy
scikit-learn
xgboost
lightgbm
matplotlib
plotly
fastf1
aiofiles
uvicorn
mysql-connector-python
sqlalchemy
python-dotenv
streamlit

# Flask and web framework dependencies
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0

# Database dependencies
mysql-connector-python==8.1.0
pymysql==1.1.0
pandas==2.1.0
sqlalchemy==2.0.21

# LangChain dependencies
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.7
langchain-openai==0.0.2
langchain-google-genai==0.0.6

# LLM API dependencies
openai==1.6.1
google-generativeai==0.3.2

# Additional utilities
logging==0.4.9.6
datetime

# Optional: For enhanced SQL parsing and analysis
sqlparse==0.4.4